+++ IsekaiMod/Content/Features/IsekaiProtagonist/OverpoweredAbility/TabletopTweaksBase/AutoBurning.cs
@@ -1,0 +1,49 @@
+﻿using IsekaiMod.Utilities;
+using Kingmaker.Blueprints;
+using Kingmaker.Blueprints.Classes;
+using Kingmaker.Designers.Mechanics.Facts;
+using Kingmaker.UnitLogic.Abilities;
+using Kingmaker.UnitLogic.Buffs.Blueprints;
+using Kingmaker.UnitLogic.FactLogic;
+using TabletopTweaks.Core.Utilities;
+using UnityEngine;
+using static IsekaiMod.Main;
+using static TabletopTweaks.Core.MechanicsChanges.MetamagicExtention;
+
+namespace IsekaiMod.Content.Features.IsekaiProtagonist.OverpoweredAbility.TabletopTweaksBase {
+
+    internal class AutoBurning {
+        private static readonly Sprite Icon_BurningSpell = BlueprintTools.GetBlueprint<BlueprintFeature>("4732a4b7b53f46848ae34a9dae66dbb2").m_Icon;
+
+        public static void Add() {
+            var AutoBurningBuff = ThingsNotHandledByTTTCore.CreateBuff("AutoBurningBuff", bp => {
+                bp.SetName(IsekaiContext, "Overpowered Ability — Auto Burning");
+                bp.SetDescription(IsekaiContext, "Every time you cast a spell, it becomes Burning, as though using the Burning Spell feat.");
+                bp.m_Icon = Icon_BurningSpell;
+                bp.AddComponent<AutoMetamagic>(c => {
+                    c.m_AllowedAbilities = AutoMetamagic.AllowedType.SpellOnly;
+                    c.Metamagic = (Metamagic)CustomMetamagic.Burning;
+                });
+                bp.Stacking = StackingType.Replace;
+                bp.IsClassFeature = true;
+                bp.m_Flags = BlueprintBuff.Flags.StayOnDeath;
+            });
+            var AutoBurningAbility = ThingsNotHandledByTTTCore.CreateActivatableAbility("AutoBurningAbility", bp => {
+                bp.SetName(IsekaiContext, "Overpowered Ability — Auto Burning");
+                bp.SetDescription(IsekaiContext, "Every time you cast a spell, it becomes Burning, as though using the Burning Spell feat.");
+                bp.m_Icon = Icon_BurningSpell;
+                bp.m_Buff = AutoBurningBuff.ToReference<BlueprintBuffReference>();
+            });
+            var AutoBurningFeature = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "AutoBurningFeature", bp => {
+                bp.SetName(IsekaiContext, "Overpowered Ability — Auto Burning");
+                bp.SetDescription(IsekaiContext, "Every time you cast a spell, it becomes Burning, as though using the Burning Spell feat.");
+                bp.m_Icon = Icon_BurningSpell;
+                bp.AddComponent<AddFacts>(c => {
+                    c.m_Facts = new BlueprintUnitFactReference[] { AutoBurningAbility.ToReference<BlueprintUnitFactReference>() };
+                });
+            });
+
+            OverpoweredAbilitySelection.AddToSelection(AutoBurningFeature);
+        }
+    }
+}